MOve type combinator narrow
