+++ lib/WorseReflection/Core/Type/AggregateType.php
@@ -88,34 +88,6 @@ abstract class AggregateType extends Type
         return $this;
     }
 
-    public function narrowTo(Type $narrowTypes): Type
-    {
-        $narrowTypes = UnionType::toUnion($narrowTypes);
-
-        if (count($narrowTypes->types) === 0) {
-            return $this;
-        }
-
-        $toRemove = [];
-        $toAdd = [];
-
-        // for each of these types
-        foreach ($this->types as $type) {
-            // check each narrowed to to see if any of these types accept the
-            // narrowed version
-            foreach ($narrowTypes->types as $narrowType) {
-                // if an existing type accepts the narrowed type, remove
-                // the existing type
-                if ($type->accepts($narrowType)->isTrue() && $type->__toString() !== $narrowType->__toString()) {
-                    $toRemove[] = $type;
-                    continue;
-                }
-            }
-        }
-
-        return $this->add($narrowTypes)->remove($this->new(...$toRemove));
-    }
-
     abstract public function new(Type ...$types): AggregateType;
 
     public function filter(): AggregateType
