+++ lib/WorseReflection/Tests/Unit/Core/Type/UnionTypeTest.php
@@ -96,116 +96,6 @@ class UnionTypeTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider provideCombine
-     * @param Type[] $types
-     */
-    public function testCombine(array $types, array $narrows, string $expected): void
-    {
-        self::assertEquals(
-            $expected,
-            TypeFactory::union(...$types)->narrowTo(TypeFactory::union(...$narrows))->__toString()
-        );
-    }
-
-    /**
-     * @return Generator<mixed>
-     */
-    public function provideCombine(): Generator
-    {
-        yield [
-            [
-                TypeFactory::string(),
-            ],
-            [
-            ],
-            'string'
-        ];
-
-        yield 'narrow to mixed' => [
-            [
-                TypeFactory::string(),
-            ],
-            [
-                TypeFactory::mixed(),
-            ],
-            'string|mixed'
-        ];
-
-        yield 'mixed narrows to int' => [
-            [
-                TypeFactory::mixed(),
-            ],
-            [
-                TypeFactory::int(),
-            ],
-            'int'
-        ];
-
-        yield 'mixed and string narrows to int' => [
-            [
-                TypeFactory::mixed(),
-                TypeFactory::string(),
-            ],
-            [
-                TypeFactory::int(),
-            ],
-            'string|int'
-        ];
-
-        yield 'empty narrow with classes' => [
-            $this->classTypes(
-                '<?php abstract class Foobar {} class Barfoo extends Foobar {}',
-                'Foobar',
-                'Barfoo',
-            ),
-            [
-            ],
-            'Foobar|Barfoo',
-        ];
-
-        yield 'narrow abstract class to concerete' => [
-            $this->classTypes(
-                '<?php abstract class Foobar {} class Barfoo extends Foobar {}',
-                'Foobar',
-                'Barfoo',
-            ),
-            [
-                TypeFactory::class('Barfoo'),
-            ],
-            'Barfoo',
-        ];
-
-        yield 'narrow abstract class to concerete with other types' => [
-            array_merge(
-                $this->classTypes(
-                    '<?php abstract class Foobar {} class Barfoo extends Foobar {}',
-                    'Foobar',
-                    'Barfoo',
-                ),
-                [
-                    TypeFactory::string(),
-                ],
-            ),
-            [
-                TypeFactory::class('Barfoo'),
-            ],
-            'Barfoo|string',
-        ];
-
-        yield 'narrow abstract with interface' => [
-            $this->classTypes(
-                '<?php interface Bar {} abstract class Foobar implements Bar {} class Barfoo extends Foobar {}',
-                'Foobar',
-                'Barfoo',
-            ),
-            [
-                TypeFactory::class('Bar'),
-            ],
-            'Foobar|Barfoo|Bar',
-        ];
-    }
-
     public function testDeduplicatesTypesOnConstruct(): void
     {
         self::assertEquals('One|Two', TypeFactory::union(
@@ -306,16 +196,5 @@ class UnionTypeTest extends TestCase
                 TypeFactory::array('string')
             )->toPhpString()
         );
-    }
-
-    /**
-     * @return Type[]
-     */
-    private function classTypes(string $string, string ...$classNames): array
-    {
-        $reflector = ReflectorBuilder::create()->addSource($string)->build();
-        return array_values(array_map(function (string $className) use ($reflector) {
-            return TypeFactory::reflectedClass($reflector, $className);
-        }, $classNames));
     }
 }
